# swagger.json
{"openapi":"3.0.0","paths":{"/auth":{"post":{"operationId":"AuthController_createUser","summary":"Cria um novo usuário","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDTO"}}}},"responses":{"201":{"description":"Usuário criado com sucesso"},"400":{"description":"Erro na solicitação"}},"tags":["Auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","summary":"Faz login de um usuário","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthUserDTO"}}}},"responses":{"200":{"description":"Login bem-sucedido"},"401":{"description":"Credenciais inválidas"}},"tags":["Auth"]}},"/auth/{id}/update-role":{"put":{"operationId":"AuthController_updateUserRole","summary":"Atualiza a função de um usuário por ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoleDTO"}}}},"responses":{"200":{"description":"Função do usuário atualizada com sucesso"},"404":{"description":"Usuário não encontrado"}},"tags":["Auth"]}},"/product-review/rate":{"post":{"operationId":"ProductReviewController_create","summary":"Avalia um produto","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductReviewDTO"}}}},"responses":{"201":{"description":"Avaliação criada com sucesso"},"400":{"description":"Erro na solicitação"}},"tags":["Product Review"],"security":[{"bearer":[]}]}},"/product-review":{"get":{"operationId":"ProductReviewController_findAll","summary":"Recupera todas as avaliações de produtos","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"Avaliações de produtos recuperadas com sucesso"},"400":{"description":"Erro na solicitação"}},"tags":["Product Review"],"security":[{"bearer":[]}]}},"/product-review/{id}":{"get":{"operationId":"ProductReviewController_findOne","summary":"Recupera uma avaliação de produto por ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Avaliação de produto recuperada com sucesso"},"404":{"description":"Avaliação de produto não encontrada"}},"tags":["Product Review"],"security":[{"bearer":[]}]},"put":{"operationId":"ProductReviewController_update","summary":"Atualiza uma avaliação de produto por ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductReviewDTO"}}}},"responses":{"200":{"description":"Avaliação de produto atualizada com sucesso"},"404":{"description":"Avaliação de produto não encontrada"}},"tags":["Product Review"],"security":[{"bearer":[]}]},"delete":{"operationId":"ProductReviewController_remove","summary":"Exclui uma avaliação de produto por ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Avaliação de produto excluída com sucesso"},"404":{"description":"Avaliação de produto não encontrada"}},"tags":["Product Review"],"security":[{"bearer":[]}]}},"/users":{"get":{"operationId":"UsersController_findAll","summary":"Recupera todos os usuários","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"Usuários recuperados com sucesso"},"400":{"description":"Erro na solicitação"}},"tags":["Users"],"security":[{"bearer":[]}]}},"/users/{id}":{"get":{"operationId":"UsersController_findOne","summary":"Recupera um usuário por ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Usuário recuperado com sucesso"},"404":{"description":"Usuário não encontrado"}},"tags":["Users"],"security":[{"bearer":[]}]},"delete":{"operationId":"UsersController_deleteUser","summary":"Exclui um usuário por ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Usuário excluído com sucesso"},"404":{"description":"Usuário não encontrado"}},"tags":["Users"],"security":[{"bearer":[]}]},"put":{"operationId":"UsersController_updateUser","summary":"Atualiza um usuário por ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDTO"}}}},"responses":{"200":{"description":"Usuário atualizado com sucesso"},"404":{"description":"Usuário não encontrado"}},"tags":["Users"],"security":[{"bearer":[]}]}},"/product":{"post":{"operationId":"ProductController_createProduct","summary":"Cria um novo produto","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDTO"}}}},"responses":{"201":{"description":"Produto criado com sucesso"},"400":{"description":"Erro na solicitação"}},"tags":["Product"],"security":[{"bearer":[]}]},"get":{"operationId":"ProductController_findAll","summary":"Recupera todos os produtos","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"Produtos recuperados com sucesso"},"400":{"description":"Erro na solicitação"}},"tags":["Product"],"security":[{"bearer":[]}]}},"/product/{id}":{"get":{"operationId":"ProductController_findOne","summary":"Recupera um produto por ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Produto recuperado com sucesso"},"404":{"description":"Produto não encontrado"}},"tags":["Product"],"security":[{"bearer":[]}]},"delete":{"operationId":"ProductController_deleteProduct","summary":"Exclui um produto por ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Produto excluído com sucesso"},"404":{"description":"Produto não encontrado"}},"tags":["Product"],"security":[{"bearer":[]}]},"put":{"operationId":"ProductController_updateUser","summary":"Atualiza um produto por ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDTO"}}}},"responses":{"200":{"description":"Produto atualizado com sucesso"},"404":{"description":"Produto não encontrado"}},"tags":["Product"],"security":[{"bearer":[]}]}},"/order":{"post":{"operationId":"OrderController_createProduct","summary":"Cria um novo pedido","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderDTO"}}}},"responses":{"201":{"description":"Pedido criado com sucesso"},"400":{"description":"Erro na solicitação"}},"tags":["Order"],"security":[{"bearer":[]}]},"get":{"operationId":"OrderController_findAll","summary":"Recupera todos os pedidos","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"Pedidos recuperados com sucesso"},"400":{"description":"Erro na solicitação"}},"tags":["Order"],"security":[{"bearer":[]}]}},"/order/{id}":{"get":{"operationId":"OrderController_findOne","summary":"Recupera um pedido por ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Pedido recuperado com sucesso"},"404":{"description":"Pedido não encontrado"}},"tags":["Order"],"security":[{"bearer":[]}]},"delete":{"operationId":"OrderController_deleteOrder","summary":"Exclui um pedido por ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Pedido excluído com sucesso"},"404":{"description":"Pedido não encontrado"}},"tags":["Order"],"security":[{"bearer":[]}]},"put":{"operationId":"OrderController_updateOrder","summary":"Atualiza um pedido por ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOrderDTO"}}}},"responses":{"200":{"description":"Pedido atualizado com sucesso"},"404":{"description":"Pedido não encontrado"}},"tags":["Order"],"security":[{"bearer":[]}]}},"/inventory/import":{"post":{"operationId":"InventoryController_upload","summary":"Importa produtos por planilha","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"201":{"description":"Produtos importados com sucesso"}},"tags":["Inventory"],"security":[{"bearer":[]}]}},"/order-item":{"post":{"operationId":"OrderItemController_create","summary":"Cria um novo item de pedido","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderItemDTO"}}}},"responses":{"201":{"description":"Item de pedido criado com sucesso"},"400":{"description":"Erro na solicitação"}},"tags":["Order Item"],"security":[{"bearer":[]}]},"get":{"operationId":"OrderItemController_findAll","summary":"Recupera todos os itens de pedido","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"Itens de pedido recuperados com sucesso"},"400":{"description":"Erro na solicitação"}},"tags":["Order Item"],"security":[{"bearer":[]}]}},"/order-item/{id}":{"get":{"operationId":"OrderItemController_findOne","summary":"Recupera um item de pedido por ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Item de pedido recuperado com sucesso"},"404":{"description":"Item de pedido não encontrado"}},"tags":["Order Item"],"security":[{"bearer":[]}]},"delete":{"operationId":"OrderItemController_deleteOrderItem","summary":"Exclui um item de pedido por ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Item de pedido excluído com sucesso"},"404":{"description":"Item de pedido não encontrado"}},"tags":["Order Item"],"security":[{"bearer":[]}]},"put":{"operationId":"OrderItemController_updateOrderItem","summary":"Atualiza um item de pedido por ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOrderItemDTO"}}}},"responses":{"200":{"description":"Item de pedido atualizado com sucesso"},"404":{"description":"Item de pedido não encontrado"}},"tags":["Order Item"],"security":[{"bearer":[]}]}}},"info":{"title":"TechSolutions","description":"API TechSolutions","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDTO":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"name":{"type":"string"},"address":{"type":"string"}},"required":["username","email","password","name","address"]},"AuthUserDTO":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"UpdateRoleDTO":{"type":"object","properties":{"role":{"type":"string","enum":["ADMIN","USER"],"description":"O novo papel do usuario","example":"ADMIN"}},"required":["role"]},"CreateProductReviewDTO":{"type":"object","properties":{"productId":{"type":"number"},"userId":{"type":"number"},"rating":{"type":"number"},"comment":{"type":"string"}},"required":["productId","userId","rating","comment"]},"UpdateProductReviewDTO":{"type":"object","properties":{"productId":{"type":"number"},"userId":{"type":"number"},"rating":{"type":"number"},"comment":{"type":"string"}},"required":["productId","userId","rating","comment"]},"UpdateUserDTO":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"name":{"type":"string"},"address":{"type":"string"}},"required":["username","email","password","name","address"]},"CreateProductDTO":{"type":"object","properties":{"price":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"stock":{"type":"number"}},"required":["price","name","description","stock"]},"UpdateProductDTO":{"type":"object","properties":{"price":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"stock":{"type":"number"}},"required":["price","name","description","stock"]},"CreateOrderDTO":{"type":"object","properties":{"userId":{"type":"number"},"status":{"type":"string","enum":["INATIVO","ATIVO","EM_ANDAMENTO","CONCLUIDO"]}},"required":["userId","status"]},"UpdateOrderDTO":{"type":"object","properties":{"userId":{"type":"number"},"status":{"type":"string","enum":["INATIVO","ATIVO","EM_ANDAMENTO","CONCLUIDO"]}},"required":["userId","status"]},"CreateOrderItemDTO":{"type":"object","properties":{"orderId":{"type":"number"},"productId":{"type":"number"},"quantity":{"type":"number"},"unitPrice":{"type":"string"}},"required":["orderId","productId","quantity","unitPrice"]},"UpdateOrderItemDTO":{"type":"object","properties":{"orderId":{"type":"number"},"productId":{"type":"number"},"quantity":{"type":"number"},"unitPrice":{"type":"string"}},"required":["orderId","productId","quantity","unitPrice"]}}}}